Helm* tests work
