+++ tests/ats/test_smoke.py
@@ -1,0 +1,41 @@
+import logging
+from typing import List
+
+import pykube
+import pytest
+from pytest_helm_charts.fixtures import Cluster, NamespaceFactoryFunc, namespace_factory  # noqa: F401
+
+# noinspection PyUnresolvedReferences
+from fixtures import (  # noqa: F401
+    flux_deployments,
+    kustomization_factory,
+    git_repository_factory,
+    helm_repository_factory,
+    helm_release_factory
+)
+
+logger = logging.getLogger(__name__)
+
+APP_DEPLOYMENT_TIMEOUT_SEC = 180
+
+
+@pytest.mark.smoke
+def test_api_working(kube_cluster: Cluster) -> None:
+    """Very minimalistic example of using the [kube_cluster](pytest_helm_charts.fixtures.kube_cluster)
+    fixture to get an instance of [Cluster](pytest_helm_charts.clusters.Cluster) under test
+    and access its [kube_client](pytest_helm_charts.clusters.Cluster.kube_client) property
+    to get access to Kubernetes API of cluster under test.
+    Please refer to [pykube](https://pykube.readthedocs.io/en/latest/api/pykube.html) to get docs
+    for [HTTPClient](https://pykube.readthedocs.io/en/latest/api/pykube.html#pykube.http.HTTPClient).
+    """
+    assert kube_cluster.kube_client is not None
+    assert len(pykube.Node.objects(kube_cluster.kube_client)) >= 1
+
+
+@pytest.mark.smoke
+def test_pods_available(
+        kube_cluster: Cluster, flux_deployments: List[pykube.Deployment]  # noqa: F811
+) -> None:
+    for d in flux_deployments:
+        assert int(d.obj["status"]["readyReplicas"]) > 0
+        logger.info(f"Deployment '{d.name}' is ready")
